server: Fix comparison at findUri
Comparing against std::string::npos requires a size_t. This mismatch
caused the comparison to always yield true, thus causing the later out of
range exception.

+++ server/rtsp.cpp
@@ -931,7 +931,7 @@ rtsp_stream *rtsp_stream::findUri(std::string uri)
 	/* for /live/ URLs, we ignore everything after the first two segments
 	 * to allow streamid and other parameters */
 	if (uri.size() > 6 && uri.compare(0, 6, "/live/") == 0) {
-		unsigned int p = uri.find('/', 6);
+		size_t p = uri.find('/', 6);
 		if (p != std::string::npos)
 			uri.erase(p);
 	}
